ref(trace-view): Move shared header component
Ripped from #25264
